Settings for default tab (global & for each user)
