+++ app/views/settings/_issue_detailed_tabs.html.erb
@@ -1,0 +1,6 @@
+<fieldset><legend><%= l(:setting_issue_history_tabs) %></legend>
+  <p>
+    <%= label_tag 'settings[:default_tab]', l(:label_issues_default_history_tab) %>
+    <%= select_tag 'settings[:default_tab]', options_for_select(%w(history_all history_comments history_private  history_activity history_time).collect{|v| [l("label_tab_#{v}"), v]}, @settings[:default_tab]) %>
+  </p>
+</fieldset>
