--- lib/issue_tabs_listener.rb
@@ -1,10 +1,0 @@
-require 'redmine'
-
-class IssueTabsViewListener < Redmine::Hook::ViewListener
-
-  # Adds javascript
-  def view_layouts_base_html_head(context)
-    javascript_include_tag('issue-history-tabs.js', :plugin => :redmine_issue_detailed_tabs_time)
-  end
-
-end
